requirements.txt
# Core frameworks
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0

# API server
fastapi>=0.115.0
uvicorn>=0.32.0
sse-starlette>=2.1.0

# Search
tavily-python>=0.5.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.9.0

# Text processing
jieba>=0.42.1
